test(core): Split trigger service tests
